Implement support for 32 bit floats

+++ test/compilation/variables/ex8.c
@@ -6,7 +6,7 @@ long* g3;
 int main() {
     int l1 = 1;
     float l2 = 3.14;
-    int l3 = 2;
+    long l3 = 2;
     g1 = &l1;
     g2 = &l2;
     g3 = &l3;
